requirements.txt
tensorflow
streamlit
numpy
PILLOW